Write final test with xfail, cleanup fixtures
